refactor : move file parsing logic to mocka-core module for ORM/ODM reuse (delete FileSourceCreator)
