+++ mocka-core/build.gradle
@@ -27,6 +27,22 @@ dependencies {
     implementation 'org.springframework:spring-context'
     testImplementation 'org.assertj:assertj-core:3.27.7'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
+
+    // parser
+    // XML
+    implementation 'jakarta.xml.bind:jakarta.xml.bind-api:4.0.4'
+    implementation 'org.glassfish.jaxb:jaxb-runtime:4.0.6'
+
+    // JSON
+    implementation 'com.fasterxml.jackson.core:jackson-databind:2.20.0'
+
+    // CSV
+    implementation 'org.apache.commons:commons-csv:1.14.1'
+
+    // XLSX
+    implementation 'org.apache.poi:poi-ooxml:5.4.1'
+
+    implementation 'org.yaml:snakeyaml:2.2'
 }
 
 // generateGrammarSource settings
