+++ mocka-orm/src/test/java/mocka/orm/entity/filesource/yaml/FileSourceYamlInvalidPathFailEntity.java
@@ -1,7 +1,7 @@
 package mocka.orm.entity.filesource.yaml;
 
 
-import mocka.orm.annotation.FileSource;
+import mocka.core.annotation.FileSource;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
