+++ mocka-orm/src/test/java/mocka/orm/entity/valuesource/ValueSourceKeyFirstSuccess.java
@@ -2,7 +2,7 @@ package mocka.orm.entity.valuesource;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
-import mocka.orm.annotation.ValueSource;
+import mocka.core.annotation.ValueSource;
 
 @Entity
 public class ValueSourceKeyFirstSuccess {
